feat: ajustes nas nomenclaturas e nas rotas
