--- server/routes/logged.js
@@ -1,11 +1,0 @@
-var express = require('express');
-var router = express.Router();
-const controller = require('../controllers/loggedController')
-
-
-router.get('/calendario', controller.calendario)
-router.get('/criar-agenda', controller.criar)
-
-
-
-module.exports = router;
